+++ packages/atproto/lib/src/atproto.dart
@@ -66,6 +66,9 @@ sealed class ATProto {
         ),
       );
 
+  /// Returns the merged headers with global headers and auth header.
+  Map<String, String> get headers;
+
   /// Returns the current session.
   ///
   /// Set only if an instance of this object was created in
@@ -160,6 +163,9 @@ final class _ATProto implements ATProto {
         label = LabelService(ctx),
         temp = TempService(ctx),
         _ctx = ctx;
+
+  @override
+  Map<String, String> get headers => _ctx.headers;
 
   @override
   core.Session? get session => _ctx.session;