+++ packages/atproto_core/test/src/clients/service_context_test.dart
@@ -1,0 +1,47 @@
+// Copyright 2024 Shinya Kato. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+// 📦 Package imports:
+import 'package:test/test.dart';
+
+// 🌎 Project imports:
+import 'package:atproto_core/src/clients/service_context.dart';
+import 'package:atproto_core/src/sessions/session.dart';
+
+void main() {
+  group('.headers', () {
+    test('case1', () {
+      final context = ServiceContext();
+
+      expect(context.headers, const <String, String>{});
+    });
+
+    test('case2', () {
+      final expected = const <String, String>{'atproto-test': '1234'};
+
+      final context = ServiceContext(headers: expected);
+
+      expect(context.headers, expected);
+    });
+
+    test('case3', () {
+      final headers = const <String, String>{'atproto-test': '1234'};
+
+      final context = ServiceContext(
+        headers: headers,
+        session: Session(
+          did: 'did:plc:iijrtk7ocored6zuziwmqq3c',
+          handle: 'shinyakato.dev',
+          accessJwt: '1234',
+          refreshJwt: '1234',
+        ),
+      );
+
+      expect(context.headers, const <String, String>{
+        'Authorization': 'Bearer 1234',
+        'atproto-test': '1234',
+      });
+    });
+  });
+}